Extract karva_coverage crate; add executable-line baseline test
