+++ Cargo.toml
@@ -17,6 +17,7 @@ karva_cache = { path = "crates/karva_cache" }
 karva_cli = { path = "crates/karva_cli" }
 karva_collector = { path = "crates/karva_collector" }
 karva_combine = { path = "crates/karva_combine" }
+karva_coverage = { path = "crates/karva_coverage" }
 karva_diagnostic = { path = "crates/karva_diagnostic" }
 karva_logging = { path = "crates/karva_logging" }
 karva_macros = { path = "crates/karva_macros" }
