+++ crates/karva/Cargo.toml
@@ -19,6 +19,7 @@ path = "src/main.rs"
 karva_cache = { workspace = true }
 karva_cli = { workspace = true }
 karva_collector = { workspace = true }
+karva_coverage = { workspace = true }
 karva_logging = { workspace = true }
 karva_metadata = { workspace = true }
 karva_project = { workspace = true }
