+++ crates/karva_test_semantic/Cargo.toml
@@ -13,6 +13,7 @@ license = { workspace = true }
 [dependencies]
 camino = { workspace = true }
 karva_collector = { workspace = true }
+karva_coverage = { workspace = true }
 karva_diagnostic = { workspace = true, features = ["traceback"] }
 karva_metadata = { workspace = true }
 karva_project = { workspace = true }
@@ -26,9 +27,6 @@ ruff_macros = { workspace = true }
 ruff_python_ast = { workspace = true }
 ruff_python_parser = { workspace = true }
 ruff_source_file = { workspace = true }
-ruff_text_size = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
 tempfile = { workspace = true }
 tracing = { workspace = true }
